+++ packages/common/src/operations/index.ts
@@ -1,2 +1,3 @@
 export { default as DeferredOperationBuilder } from "./deferredOperationBuilder";
 export { default as HttpOperationBuilder } from "./httpOperationBuilder";
+export { default as HttpOperationBuilderFactory } from "./httpOperationBuilderFactory";