+++ packages/common/src/types.ts
@@ -1,3 +1,10 @@
+import { BaseError } from "./errors";
 import { LoggerFactoryEnvironmentKeys } from "./logger";
 
 export type CommonModuleEnvironmentKeys = LoggerFactoryEnvironmentKeys;
+
+export interface OperationResult<DataType> {
+	data: DataType | null;
+	error: BaseError | null;
+	success: boolean;
+}