+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.2-rc-2] - 06/20/2026
+## [1.1.2-rc-2] - 06/27/2026
 ### Changed
+* Updating handling of http operation
 * Updating to pnpm v11.8.0
 * Updating the environmental variables