+++ packages/archives/CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.1-rc-1] - 06/20/2026
+## [1.0.1-rc-1] - 06/27/2026
 ### Added
 * Adding typescript build script
 * Adding a `ArchivesModule` and `ContactModule`
 
 ### Changed
+* Properly typing `ContactHttpClient`
 * Consuming changes to `EnvironmentRegistry`
 * Updating to importing `withRateLimit` preHook through the new name: `withRateLimitPreHook`
 * Updating `tsconfig.ts` to properly resolve other packages within this repo